requirements.txt
Cython
prophet
seaborn
matplotlib
scikit-learn
jupyterlab
notebook
pandas-profiling
statsmodels
scalecast
pmdarima